code refactoring to allow simulation
