+++ mdp_plan_exec/launch/mdp_plan_exec_extended.launch
@@ -14,5 +14,5 @@
     <node pkg="prism_strands" type="prism" name="prism_executor"  output="screen" args="$(arg executor_port) $(arg executor_dir) $(arg prism_model_file) -javamaxmem $(arg javamaxmem)" respawn="true"/>
 
     <node pkg="mdp_plan_exec" type="mdp_task_guarantees_estimator.py" name="mdp_task_guarantees_estimator"  output="screen" args="$(arg estimator_port) $(arg estimator_dir) $(arg prism_model_file)"/>
-    <node pkg="mdp_plan_exec" type="mdp_policy_executor_extended.py" name="mdp_policy_executor_extended"  output="screen" args="$(arg executor_port) $(arg executor_dir) $(arg prism_model_file)"/>
+    <node pkg="mdp_plan_exec" type="robot_mdp_exec.py" name="robot_mdp_policy_executor"  output="screen" args="$(arg executor_port) $(arg executor_dir) $(arg prism_model_file)"/>
 </launch>
